Replace usage of `yes,no` enum types by `n,y`
Well, these enum types were introduced quite years ago, and it was fine so far,
as we were creating the queries manually with ipl-sql. However, now we want to
switch everything to ipl-orm, which makes it more cumbersome to use, e.g, the
`BoolCast` behavior on the fly, as it expects the booleums to contain `n,y` and
not `yes,no`. Due to the need to use this behavior, it is much better to rename these
enums beforehand than to introduce a separate behavior.