Don't override scheduled jobs with the new ones
Before Icinga PHP Library `<=0.13.0`, the scheduler was using the PHP
standard `SplObjectStorage`, which automatically generates a unique hash
whenever adding a new task using `attach()`. This hash was generated
internally with `spl_object_hash()`. As a result, when we try to remove
a job from the scheduler that was scheduled e.g `5m` ago (since this is
the default daemon config reload interval), the scheduler crashes with
job `foo` not scheduled.