CheckCommand: Drop superfluous `issuer_certificate` joins
It doesn't make sense to directly join on the `issuer` here at all. The
join was utilized to determine if the certificate with `link.order = 0`
is self-seigned, but this often results in incorrect conclusions, as
most of the issuers are self-signed. Instead, the `self_signed` column
available for each certificate should be used for this purpose.